requirements.txt
pytorch
matplotlib